put user/instance related methods under its own struct
