+++ examples/dynamic_scene/src/main.rs
@@ -243,6 +243,6 @@ fn main() {
                 }
             }
         }
-        time += 1.0;
+        time += 0.2;
     });
 }
