+++ src/lib.rs
@@ -18,7 +18,6 @@ mod callback;
 mod device;
 mod error;
 mod geometry;
-mod instance;
 mod intersect_context;
 mod ray;
 mod scene;
@@ -33,7 +32,6 @@ pub use buffer::*;
 pub use device::*;
 pub use error::*;
 pub use geometry::*;
-pub use instance::*;
 pub use intersect_context::*;
 pub use ray::*;
 pub use scene::*;
